bugfixes, credits, tutorial, gui, animations

+++ BlobGame/Game/GUI/GUILabel.cs
@@ -5,7 +5,7 @@ using System.Numerics;
 namespace BlobGame.Game.Gui;
 internal sealed class GuiLabel {
     private string Text { get; }
-    private int FontSize { get; }
+    private float FontSize { get; }
 
     private Vector2 TextPosition { get; }
 
@@ -20,7 +20,7 @@ internal sealed class GuiLabel {
         }
 
         Text = text;
-        FontSize = (int)(h * 0.6f);
+        FontSize = h * 0.6f;
         Vector2 textSize = Raylib.MeasureTextEx(Renderer.Font.Resource, text, FontSize, FontSize / 16f);
         TextPosition = new Vector2(x + w / 2 - textSize.X / 1.75f, y + h / 2 - FontSize / 2);
     }
